Fix [UI] App crashes on opening Function/Models details
