+++ src/common/TabsSlider/TabsSlider.js
@@ -25,6 +25,7 @@ import classnames from 'classnames'
 import { Tip } from 'igz-controls/components'
 
 import { SLIDER_STYLE_1, SLIDER_STYLE_2, SLIDER_TABS } from '../../types'
+import { generateUrlFromRouterPath  } from '../../utils/link-helper.util'
 
 import { ReactComponent as Arrow } from 'igz-controls/images/arrow.svg'
 
@@ -116,8 +117,8 @@ const TabsSlider = ({
     } else if (
       tabsRef.current?.scrollWidth <
       tabsWrapperRef.current?.offsetWidth / menuOffsetHalfWidth +
-        selectedTabNode?.offsetLeft +
-        selectedTabNode?.offsetWidth
+      selectedTabNode?.offsetLeft +
+      selectedTabNode?.offsetWidth
     ) {
       setScrolledWidth(tabsRef.current?.scrollWidth - tabsWrapperRef.current?.offsetWidth)
       setRightArrowDisabled(true)
@@ -188,11 +189,7 @@ const TabsSlider = ({
                 <Link
                   className={tabClassName}
                   data-tab={tab.id}
-                  to={
-                    new URL(
-                      `${location.pathname?.replace(/^$|([^/]+$)/, tab.id)}${location.search ?? ''}`
-                    )
-                  }
+                  to={generateUrlFromRouterPath (`${location.pathname?.replace(/^$|([^/]+$)/, tab.id)}${location.search ?? ''}`)}
                   onClick={() => onSelectTab(tab)}
                   key={tab.id}
                 >
