+++ src/utils/createArtifactsContent.js
@@ -32,8 +32,7 @@ import { parseKeyValues } from './object'
 import { formatDatetime } from './datetime'
 import prettyBytes from 'pretty-bytes'
 import { parseUri } from './parseUri'
-import { generateFunctionDetailsLink } from './generateFunctionDetailsLink'
-import { generateLinkToDetailsPanel } from './generateLinkToDetailsPanel'
+import { generateFunctionDetailsLink, generateLinkToDetailsPanel } from './link-helper.util'
 import { validateArguments } from './validateArguments'
 // import { roundFloats } from './roundFloats'
 import TableProducerCell from '../elements/TableProducerCell/TableProducerCell'
