+++ src/utils/createJobsContent.js
@@ -32,8 +32,7 @@ import {
 } from '../components/Workflow/workflow.util'
 import measureTime from './measureTime'
 import { formatDatetime } from './datetime'
-import { generateFunctionDetailsLink } from './generateFunctionDetailsLink'
-import { generateLinkToDetailsPanel } from './generateLinkToDetailsPanel'
+import { generateFunctionDetailsLink, generateLinkToDetailsPanel } from './link-helper.util'
 import { getJobIdentifier, getWorkflowJobIdentifier } from './getUniqueIdentifier'
 import { parseKeyValues } from './object'
 import { validateArguments } from './validateArguments'
